fix: add missing Admin namespace files for CI build
Move Admin functionality from Server to Core/Postgres architecture:
- Add ITableDiscoveryService interface in Core.Features.Admin.Abstractions
- Add TableDiscoveryModels in Core.Features.Admin.Domain
- Move PostgreSqlTableDiscoveryService to Postgres.Features.Admin
- Update service registrations and endpoint references
- Add TableDiscoveryJsonContext for JSON serialization

This resolves the CI build failures caused by missing Admin namespace references.
