+++ src/Honua.Server/Features/Admin/AdminEndpoints.cs
@@ -2,9 +2,10 @@
 // Licensed under the Elastic License 2.0. See LICENSE in the project root.
 
 using System.Text.Json;
+using Honua.Core.Features.Admin.Abstractions;
+using Honua.Core.Features.Admin.Domain;
 using Honua.Core.Features.Infrastructure.Abstractions;
 using Honua.Server.Features.Admin.Models;
-using Honua.Server.Features.Admin.Services;
 
 namespace Honua.Server.Features.Admin;
